LayoutTests/fullscreen/fullscreen-containing-block-change-expected.txt

Test passes if the lightgreen area fills the screen after going fullscreen.

Go fullscreen
EXPECTED (target.offsetWidth > 400 == 'true') OK
EXPECTED (target.offsetHeight > 400 == 'true') OK
END OF TEST
